✨ Display waiting opponent screen
